pyproject.toml
[project]
name = "wagtail-starling"
version = "0.1.0"
description = "Reusable Wagtail components for multilingual content management"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "wagtail>=6.4.1",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-django>=4.5.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"
